feat!(applet): rename Home to About
