+++ packages/applet/src/modules/custom-inspector/components/About.vue
