+++ packages/applet/src/modules/pinia/components/About.vue
