+++ packages/applet/src/modules/pinia/index.vue
@@ -1,16 +1,10 @@
 <script setup lang="ts">
-import Home from './components/Home.vue'
+import About from './components/About.vue'
 import Store from './components/store/Index.vue'
 import Timeline from './components/timeline/Index.vue'
 import { registerVirtualRouter } from '~/composables/virtual-router'
 
 const { VirtualRouterView } = registerVirtualRouter([
-  {
-    path: '/',
-    name: 'Home',
-    component: Home,
-    icon: 'i-logos-pinia',
-  },
   {
     path: '/store',
     name: 'Store',
@@ -22,6 +16,12 @@ const { VirtualRouterView } = registerVirtualRouter([
     name: 'Timeline',
     component: Timeline,
     icon: 'i-mdi:timeline-clock-outline',
+  },
+  {
+    path: '/',
+    name: 'About',
+    component: About,
+    icon: 'i-logos-pinia',
   },
 ], {
   defaultRoutePath: '/store',
